+++ modules/Core/src/Registration/Feature.cpp
@@ -85,31 +85,31 @@ std::shared_ptr<Feature> ComputeSPFHFeature(const PointCloud &input,
         const KDTreeFlann &kdtree, const KDTreeSearchParam &search_param)
 {
     auto feature = std::make_shared<Feature>();
-    feature->Resize(33, (int)input.points_.size());
+    feature->Resize(33, (int32_t)input.points_.size());
 #ifdef _OPENMP
 #pragma omp parallel for schedule(static)
 #endif
-    for (int i = 0; i < (int)input.points_.size(); i++) {
+    for (int32_t i = 0; i < (int32_t)input.points_.size(); i++) {
         const auto &point = input.points_[i];
         const auto &normal = input.normals_[i];
-        std::vector<int> indices;
+        std::vector<int32_t> indices;
         std::vector<double> distance2;
         if (kdtree.Search(point, search_param, indices, distance2) > 1) {
             // only compute SPFH feature when a point has neighbors
             double hist_incr = 100.0 / (double)(indices.size() - 1);
-            for (size_t k = 1; k < indices.size(); k++) {
+            for (uint32_t k = 1; k < indices.size(); k++) {
                 // skip the point itself, compute histogram
                 auto pf = ComputePairFeatures(point, normal,
                         input.points_[indices[k]], input.normals_[indices[k]]);
-                int h_index = (int)(floor(11 * (pf(0) + M_PI) / (2.0 * M_PI)));
+                int32_t h_index = (int32_t)(floor(11 * (pf(0) + M_PI) / (2.0 * M_PI)));
                 if (h_index < 0) h_index = 0;
                 if (h_index >= 11) h_index = 10;
                 feature->data_(h_index, i) += hist_incr;
-                h_index = (int)(floor(11 * (pf(1) + 1.0) * 0.5));
+                h_index = (int32_t)(floor(11 * (pf(1) + 1.0) * 0.5));
                 if (h_index < 0) h_index = 0;
                 if (h_index >= 11) h_index = 10;
                 feature->data_(h_index + 11, i) += hist_incr;
-                h_index = (int)(floor(11 * (pf(2) + 1.0) * 0.5));
+                h_index = (int32_t)(floor(11 * (pf(2) + 1.0) * 0.5));
                 if (h_index < 0) h_index = 0;
                 if (h_index >= 11) h_index = 10;
                 feature->data_(h_index + 22, i) += hist_incr;
@@ -125,7 +125,7 @@ std::shared_ptr<Feature> ComputeFPFHFeature(const PointCloud &input,
         const KDTreeSearchParam &search_param/* = KDTreeSearchParamKNN()*/)
 {
     auto feature = std::make_shared<Feature>();
-    feature->Resize(33, (int)input.points_.size());
+    feature->Resize(33, (int32_t)input.points_.size());
     if (input.HasNormals() == false) {
         PrintDebug("[ComputeFPFHFeature] Failed because input point cloud has no normal.\n");
         return feature;
@@ -135,26 +135,26 @@ std::shared_ptr<Feature> ComputeFPFHFeature(const PointCloud &input,
 #ifdef _OPENMP
 #pragma omp parallel for schedule(static)
 #endif
-    for (int i = 0; i < (int)input.points_.size(); i++) {
+    for (int32_t i = 0; i < (int32_t)input.points_.size(); i++) {
         const auto &point = input.points_[i];
-        std::vector<int> indices;
+        std::vector<int32_t> indices;
         std::vector<double> distance2;
         if (kdtree.Search(point, search_param, indices, distance2) > 1) {
             double sum[3] = {0.0, 0.0, 0.0};
-            for (size_t k = 1; k < indices.size(); k++) {
+            for (uint32_t k = 1; k < indices.size(); k++) {
                 // skip the point itself
                 double dist = distance2[k];
                 if (dist == 0.0)
                     continue;
-                for (int j = 0; j < 33; j++) {
+                for (int32_t j = 0; j < 33; j++) {
                     double val = spfh->data_(j, indices[k]) / dist;
                     sum[j / 11] += val;
                     feature->data_(j, i) += val;
                 }
             }
-            for (int j = 0; j < 3; j++)
+            for (int32_t j = 0; j < 3; j++)
                 if (sum[j] != 0.0) sum[j] = 100.0 / sum[j];
-            for (int j = 0; j < 33; j++) {
+            for (int32_t j = 0; j < 33; j++) {
                 feature->data_(j, i) *= sum[j / 11];
                 // The commented line is the fpfh function in the paper.
                 // But according to PCL implementation, it is skipped.
